requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
tensorflow
scikit-learn
opencv-python
scikit-image
matplotlib
numpy
scipy
seaborn
flake8
python-dotenv>=0.5.1
